Modification Splash screen activity
