--- app/src/main/java/dev/munderstand/epsi/SplashActivity.kt
@@ -1,21 +1,0 @@
-package dev.munderstand.epsi
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-
-class SplashActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
-        Handler(Looper.getMainLooper()).postDelayed(Runnable {
-            // ADD LOGIC FOR ACCOUNT OR NOT
-            val newIntent = Intent(application, SignInActivity::class.java)
-            startActivity(newIntent)
-            finish()
-        }, 2000)
-
-    }
-}
